fix: Separate client and server components to resolve generateMetadata error
- Split projects page into server component (page.tx) and client component (ProjectsClient.tsx)
- Move all interactive functionality to client component
- Keep generateMetadata in server component for proper SEO
- Fix Next.js app directory architecture compliance
- Maintain all existing functionality while resolving build errors
